Initial commit.
Known issues:
We iterate every media type and say "hey, can you handle this?". If the
request has a "*/*" every media type says yes. We should move this logic
to the Media class and just have it delegate to MediaType for rendering.

Iterate all `types` first since those are most important.

Compile `accept` and ask negotiator if it likes any of these.